requirements.txt
flask
flask-session
python-dotenv
supabase
requests
